implemented latency in serial driver and sensor

+++ dsm/modules/a2d_driver.h
@@ -80,7 +80,7 @@ typedef struct
 	int	offset[MAXA2DS];	// Offset flags
 	US	master;		// Designates master A/D
 	US	filter[CONFBLOCKS*CONFBLLEN+1];	// Filter data
-	int	latencyUsecs;	// buffer latency in micro-seconds
+	long	latencyUsecs;	// buffer latency in micro-seconds
 } A2D_SET;
 
 typedef struct
